+++ src/favorite/favorites.ts
@@ -24,12 +24,12 @@ export default class Favorites {
         if (item && item.is_favorite) evo.is_favorite = true
       })
     }
-    nb_favorite.update(n => count_nb_favorite)
+    nb_favorite.set(count_nb_favorite)
   }
   static clear() {
     this.favorites = []
     this.save()
-    nb_favorite.update(n => 0)
+    nb_favorite.set(0)
   }
   static save() {
     db.browser.set(this.db_key, this.favorites)